Various updates to parsing, still a work in progress
The primary change is to support the type annotations syntax. Currently
type annotations are ignored for values, and only exposed for nodes. We
don't actually do anything with them though.

I also renamed a few things, and introduced a new NodeList class. This
is meant to be the intermediary used for consistent handling of lists of
nodes, because children are just lists of nodes, same as the top-level
document.

I also removed the last of the backwards-compatible "convert to string"
handling in the structure classes.