Much refactoring to improve correctness
Some of this involved updating the grammar, and some of this involved
rewriting how individual values are encoded and decoded.

Bools and nulls can now be manually parsed to accommodate arbitrary type
annotations on these values. A consequence of this change is that any
code that overrides a particular type parser must call the default
parser for that type when no value is recognised.